move logger package into the core directory
