+++ core/cmd/enclave_sgx.go
@@ -2,7 +2,7 @@
 
 package cmd
 
-import "github.com/smartcontractkit/chainlink/logger"
+import "github.com/smartcontractkit/chainlink/core/logger"
 
 // InitEnclave initialized the SGX enclave for use by adapters
 func InitEnclave() error {
