+++ core/cmd/prompter.go
@@ -7,7 +7,7 @@ import (
 	"os/signal"
 	"strings"
 
-	"github.com/smartcontractkit/chainlink/logger"
+	"github.com/smartcontractkit/chainlink/core/logger"
 	"golang.org/x/crypto/ssh/terminal"
 )
