+++ core/logger/prettyconsole_test.go
