+++ services/application.go
@@ -8,7 +8,7 @@ import (
 	"syscall"
 
 	"github.com/gobuffalo/packr"
-	"github.com/smartcontractkit/chainlink/logger"
+	"github.com/smartcontractkit/chainlink/core/logger"
 	"github.com/smartcontractkit/chainlink/store"
 	"github.com/smartcontractkit/chainlink/store/models"
 	"github.com/smartcontractkit/chainlink/utils"
