+++ services/reaper.go
@@ -3,7 +3,7 @@ package services
 import (
 	"time"
 
-	"github.com/smartcontractkit/chainlink/logger"
+	"github.com/smartcontractkit/chainlink/core/logger"
 	"github.com/smartcontractkit/chainlink/store"
 )
