+++ services/scheduler.go
@@ -6,7 +6,7 @@ import (
 	"time"
 
 	"github.com/mrwonko/cron"
-	"github.com/smartcontractkit/chainlink/logger"
+	"github.com/smartcontractkit/chainlink/core/logger"
 	"github.com/smartcontractkit/chainlink/store"
 	"github.com/smartcontractkit/chainlink/store/models"
 	"github.com/smartcontractkit/chainlink/utils"
